+++ src/mainQt3D.cpp
@@ -8,7 +8,7 @@
 #include "rns.h"
 
 
-static float rnsversion = 0.8;
+static double rnsversion = 0.8;
 
 int main(int argc, char *argv[])
 {
